# web/openapi.yaml
openapi: 3.0.3
info:
  title: Fikua Attestation Registry
  description: >
    Catalogue of Attestations for the Fikua Lab EUDI Wallet ecosystem, per
    ARF 3.0 and ETSI TS 119 472. Read-only JSON API consumed by
    fikua-lab-issuer and fikua-lab-verifier.
  version: "0.1.0"
  license:
    name: Proprietary
paths:
  /health:
    get:
      summary: Health check
      operationId: health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/v1/schemes:
    get:
      summary: List all attestation definitions
      operationId: listSchemes
      responses:
        "200":
          description: All registered attestation definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttestationDefinition"
  /api/v1/schemes/{id}:
    get:
      summary: Get one attestation definition by scheme id
      operationId: getScheme
      parameters:
        - name: id
          in: path
          required: true
          description: >
            The attestation scheme id (e.g. "urn:eudi:pid:1" or
            "eu.europa.ec.eudi.pid.1"). May contain colons.
          schema:
            type: string
      responses:
        "200":
          description: The attestation definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttestationDefinition"
        "404":
          description: No definition registered under that scheme id
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/v1/schemes/{id}/{version}/type-metadata:
    get:
      summary: >
        Get the SD-JWT VC Type Metadata Document for a scheme's dc+sd-jwt
        format at an exact version, per the SD-JWT VC specification's
        "Registry" retrieval method (this registry IS that registry — see
        https://www.ietf.org/archive/id/draft-ietf-oauth-sd-jwt-vc-latest.html#name-sd-jwt-vc-type-metadata).
        Immutable and version-addressed (FR-18/19/20): there is no
        unversioned alias. version must match the scheme's current
        `version` field (GET /api/v1/schemes/{id}); this registry does not
        yet keep prior versions' content addressable. Only defined for
        schemes that support dc+sd-jwt; 404 otherwise (e.g. the PID mdoc
        scheme).
      operationId: getTypeMetadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          description: Must equal the scheme's current version exactly.
          schema:
            type: string
      responses:
        "200":
          description: SD-JWT VC Type Metadata Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TypeMetadata"
        "404":
          description: >
            No such scheme, the scheme has no dc+sd-jwt format, or version
            does not match the scheme's current version
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
components:
  schemas:
    AttestationDefinition:
      type: object
      properties:
        rulebook:
          $ref: "#/components/schemas/AttestationRulebook"
        scheme:
          $ref: "#/components/schemas/AttestationScheme"
    AttestationRulebook:
      type: object
      description: Human-readable Attestation Rulebook metadata (ARF 3.0 / TS11 §4.2).
      properties:
        attestationType:
          type: string
        status:
          type: string
        version:
          type: string
        contact:
          type: string
        legalBasis:
          type: array
          items:
            type: string
        category:
          type: string
          enum:
            - urn:etsi:esi:eaa:eu:qualified
            - urn:etsi:esi:eaa:eu:pub
            - "eaa:eu:non-qualified"
        documentUri:
          type: string
        revocation:
          type: string
          description: Rulebook template §6 (SHALL for every Rulebook).
          enum:
            - not_applicable_short_lived
            - attestation_status_list
            - attestation_revocation_list
        maxValiditySeconds:
          type: integer
          description: >
            Maximum technical validity (issued-token exp-iat/nbf) an issuer
            may set. Required, and must be ≤86400, when revocation is
            not_applicable_short_lived (ARF 3.0 Topic 7 VCR_01b: the
            short-lived basis for skipping revocation only holds for a
            validity period of 24 hours or less, per ETSI EN 319 411-1
            REV-6.2.4-03A).
        revocationListUrl:
          type: string
          description: Where to retrieve the ASL/ARL. Empty when revocation is not_applicable_short_lived.
        trustAnchorDescription:
          type: string
          description: Rulebook template §5 prose (SHOULD for non-qualified EAA, ARB_26).
        proximityUseCaseAnalysis:
          type: string
          description: >
            ARB_02's required documented analysis of whether this
            attestation type must be presentable in proximity without
            internet connectivity — if so, mso_mdoc must be a supported
            format, since OpenID4VP (the only presentation protocol for
            dc+sd-jwt/W3C VC) requires connectivity.
        dataSeparationStatement:
          type: string
          description: >
            eIDAS 2.0 Art. 45h's SHALL requirement (applies to non-qualified
            EAA providers too): personal data used to issue this
            attestation is not combined with personal data from any other
            service the issuer or its commercial partners offer.
    AttestationScheme:
      type: object
      description: Machine-readable attestation schema (ARF 3.0 / TS11 §4.3.1 SchemaMeta).
      properties:
        catalogueId:
          type: string
          description: >
            TS11 §4.3.1 SchemaMeta.id: an opaque UUID assigned by the
            catalogue provider. `id` (below) is this registry's own
            human-readable lookup/URL identifier instead — see
            docs/compliance/*.md for why they're kept separate.
        id:
          type: string
        version:
          type: string
        rulebookUri:
          type: string
        trustedAuthorities:
          type: array
          items:
            $ref: "#/components/schemas/TrustAuthority"
        attestationLoS:
          type: string
          enum: [iso_18045_high, iso_18045_moderate, iso_18045_enhanced-basic, iso_18045_basic]
        bindingType:
          type: string
          enum: [claim, key, biometric, none]
        supportedFormats:
          type: array
          items:
            type: string
            enum: ["dc+sd-jwt", mso_mdoc]
        schemaURIs:
          type: array
          description: >
            TS11 §4.3.2 Schema sub-class: {formatIdentifier, uri} pointers to
            the format-native schema document for each supported format
            (SD-JWT VC Type Metadata for dc+sd-jwt; mso_mdoc has no
            generator yet, so its uri is empty). Computed per-request from
            the request's host/basePath, not stored. Content lives at uri,
            not here — contrast with `schemas` below, this registry's own
            richer internal representation.
          items:
            $ref: "#/components/schemas/SchemaURI"
        schemas:
          type: array
          description: >
            This registry's own internal claim content per format — not the
            TS11 Schema sub-class itself (see schemaURIs above for that).
            Kept in the API response because fikua-lab-issuer/verifier
            consume it directly.
          items:
            $ref: "#/components/schemas/FormatSchema"
    SchemaURI:
      type: object
      description: TS11 §4.3.2 Schema sub-class — a pointer, not a container.
      properties:
        formatIdentifier:
          type: string
          enum: ["dc+sd-jwt", mso_mdoc]
        uri:
          type: string
          description: >
            Absolute URL of the format-native schema document. Empty when no
            generator exists yet for that format (currently: mso_mdoc).
    FormatSchema:
      type: object
      description: >
        This registry's own internal representation of one issuance
        format's claim content — not the TS11 §4.3.2 Schema sub-class
        itself (see SchemaURI above for that). Source data for
        internal/sdjwtvc.FromScheme's generated Type Metadata Document.
      properties:
        format:
          type: string
          enum: ["dc+sd-jwt", mso_mdoc]
        typeIdentifier:
          type: string
          description: vct for SD-JWT VC, mdoc doctype for mso_mdoc.
        claims:
          type: array
          items:
            $ref: "#/components/schemas/ClaimDefinition"
    ClaimDefinition:
      type: object
      properties:
        dataIdentifier:
          type: string
        path:
          type: array
          items:
            type: string
        dataType:
          type: string
        presence:
          type: string
          enum: [mandatory, optional, conditional]
        namespace:
          type: string
          description: mdoc claims only.
        disclosability:
          type: string
          enum: ["MUST", "MAY", "MUST NOT"]
          description: SD-JWT VC claims only (ARB_30).
        enum:
          type: array
          items:
            type: string
          description: >
            Restricts the claim's value to this fixed set, when the
            Rulebook or an external spec mandates specific literal values
            (e.g. ETSI TS 119 472-1's fixed category-style values). Empty
            when unconstrained.
    TrustAuthority:
      type: object
      properties:
        frameworkType:
          type: string
          enum: [aki, etsi_tl, openid_federation]
        value:
          type: string
        isLoTE:
          type: boolean
    TypeMetadata:
      type: object
      description: >
        SD-JWT VC Type Metadata Document, per
        https://www.ietf.org/archive/id/draft-ietf-oauth-sd-jwt-vc-latest.html#name-type-metadata-format.
        This registry only populates vct/name/description/display/claims;
        the spec allows additional properties not listed here.
      properties:
        vct:
          type: string
        name:
          type: string
        description:
          type: string
        display:
          type: array
          items:
            type: object
            properties:
              locale:
                type: string
              name:
                type: string
              description:
                type: string
        claims:
          type: array
          items:
            $ref: "#/components/schemas/TypeMetadataClaim"
    TypeMetadataClaim:
      type: object
      description: One claim's metadata, per the spec's "Claim Metadata" section.
      properties:
        path:
          type: array
          items:
            type: string
        mandatory:
          type: boolean
        sd:
          type: string
          enum: [always, allowed, never]